@@ -4,21 +4,20 @@
   "private": true,
   "homepage": ".",
   "dependencies": {
-    "@emotion/react": "^11.8.2",
-    "@emotion/styled": "^11.8.1",
-    "@mapcomponents/react-maplibre": "^0.1.61",
-    "@material-ui/core": "^4.12.3",
-    "@mui/icons-material": "^5.6.2",
-    "@mui/material": "^5.5.0",
+    "@emotion/react": "^11.11.1",
+    "@emotion/styled": "^11.11.0",
+    "@mapcomponents/react-maplibre": "^0.1.85",
+    "@mui/icons-material": "^5.14.8",
+    "@mui/material": "^5.14.8",
     "@mui/styles": "^5.14.7",
-    "@testing-library/react": "^12.1.3",
-    "@testing-library/user-event": "^13.5.0",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.4.3",
     "buffer": "^6.0.3",
-    "deck.gl": "^8.7.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-scripts": "^5.0.0",
-    "web-vitals": "^2.1.4"
+    "deck.gl": "^8.9.27",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-scripts": "^5.0.1",
+    "web-vitals": "^3.4.0"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -34,9 +33,9 @@
   },
   "browserslist": {
     "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
+      "defaults",
+      "not op_mini all",
+      "not ie 11"
     ],
     "development": [
       "last 1 chrome version",
